Whitelist discord.gg/xbox from invite auto-warn
The Xbox server is verified and people often refer others to it for support. Unfortunately, those who are Tier 1 seem to do this more often than you'd expect, so they're often warned just for sending a link to the Xbox server. Whitelisting it would solve this problem, and it's a trusted server so there shouldn't be any concern about whitelisting it.